wip

Former-commit-id: 158d62c473a85f788777f0f15e02655291b5fcdf [formerly 158d62c473a85f788777f0f15e02655291b5fcdf [formerly f5a45cb8c0d2135e855708a134a90b552de11b02]]
Former-commit-id: 6e8a4279a773b7bea4cbc51c5ffc5887f8435fb7
Former-commit-id: 9c802a7261fae024adedf67de44a23fd140b66d5

+++ src/localization/position.h
@@ -12,7 +12,7 @@ using point3d_t = struct Point3d {
 
 using tag_detection_t = struct TagDetection {
   int tag_id;
-  std::array<cv::Point2f, 4> corners;  // Image coordinates of tag corners
+  std::array<cv::Point2d, 4> corners;  // Image coordinates of tag corners
   double timestamp;
   double confidence;
 };
